+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## New in git-machete 3.11.2
 
+- fixed: `git machete` now correctly infers remote for fetching of branch when the branch is found in more than one remote
+
 ## New in git-machete 3.11.1
 
 - fixed: release to Snap Store